refactor(cli): extract convert subcommand to internal/convertcli
